DADAO-tests: rewrite llvm-testsuite run script in python
A shell script to build and run all test cases was written because
"make all" stops at the first target it fails to build. That shell
script is replaced by a python script because in shell scripts it
is tricky to keep trailing newlines ('\n') in command substitution.
Also, it will be easier to support building and runnning test cases
in parallel using python.

Signed-off-by: Guorui Wen <grwen@stu.pku.edu.cn>